backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.0.post1
motor==3.3.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
Pillow==10.2.0
python-jose==3.3.0
passlib==1.7.4
email-validator==2.1.0.post1
python-dotenv==1.0.0